refactor io
