+++ ramannoodle/io/vasp/outcar.py
@@ -1,20 +1,21 @@
-"""Utilities for VASP.
-
-One should be careful about running these functions, as they will often only partially
-read through a given file. Each function must be run in a specific order.
-
-"""
+"""Functions for interacting with VASP OUTCAR files."""
 
 from typing import TextIO
+from pathlib import Path
 import numpy as np
 from numpy.typing import NDArray
 
-from ramannoodle.io.io_utils import _skip_file_until_line_contains
+
+from ramannoodle.io.io_utils import _skip_file_until_line_contains, pathify
 from ramannoodle.exceptions import InvalidFileException, NoMatchingLineFoundException
-from ramannoodle.globals import ATOMIC_NUMBERS
+from ramannoodle.globals import ATOMIC_WEIGHTS, ATOMIC_NUMBERS
 from ramannoodle.exceptions import get_type_error
+from ramannoodle.dynamics.phonon import Phonons
+from ramannoodle.symmetry.structural import ReferenceStructure
 
 
+# Utilities for OUTCAR
+# These functions partially read files.
 def _get_atomic_symbol_from_potcar_line(line: str) -> str:
     """Extract atomic symbol from a POTCAR line in a VASP OUTCAR file.
 
@@ -34,7 +35,7 @@ def _get_atomic_symbol_from_potcar_line(line: str) -> str:
     return symbol
 
 
-def _read_atomic_symbols_from_outcar(outcar_file: TextIO) -> list[str]:
+def _read_atomic_symbols(outcar_file: TextIO) -> list[str]:
     """Read atomic symbols from a VASP OUTCAR file.
 
     Raises
@@ -77,9 +78,7 @@ def _read_atomic_symbols_from_outcar(outcar_file: TextIO) -> list[str]:
         ) from exc
 
 
-def _read_eigenvector_from_outcar(
-    outcar_file: TextIO, num_atoms: int
-) -> NDArray[np.float64]:
+def _read_eigenvector(outcar_file: TextIO, num_atoms: int) -> NDArray[np.float64]:
     """Read the next available phonon eigenvector from a VASP OUTCAR file.
 
     Raises
@@ -99,7 +98,7 @@ def _read_eigenvector_from_outcar(
         raise InvalidFileException(f"eigenvector could not be parsed: {line}") from exc
 
 
-def _read_cartesian_positions_from_outcar(
+def _read_cartesian_positions(
     outcar_file: TextIO, num_atoms: int
 ) -> NDArray[np.float64]:
     """Read atomic cartesian positions from a VASP OUTCAR file.
@@ -127,7 +126,7 @@ def _read_cartesian_positions_from_outcar(
     return np.array(cartesian_coordinates)
 
 
-def _read_fractional_positions_from_outcar(
+def _read_fractional_positions(
     outcar_file: TextIO, num_atoms: int
 ) -> NDArray[np.float64]:
     """Read atomic fractional positions from a VASP OUTCAR file.
@@ -155,7 +154,7 @@ def _read_fractional_positions_from_outcar(
     return np.array(fractional_coordinates)
 
 
-def _read_polarizability_from_outcar(outcar_file: TextIO) -> NDArray[np.float64]:
+def _read_polarizability(outcar_file: TextIO) -> NDArray[np.float64]:
     """Read polarizability from a VASP OUTCAR file.
 
     In actuality, we read the macroscopic dielectric tensor including local field
@@ -202,7 +201,7 @@ def _get_lattice_vector_from_outcar_line(line: str) -> NDArray[np.float64]:
         raise ValueError("line does not have the expected format") from exc
 
 
-def _read_lattice_from_outcar(outcar_file: TextIO) -> NDArray[np.float64]:
+def _read_lattice(outcar_file: TextIO) -> NDArray[np.float64]:
     """Read all three lattice vectors (in angstroms) from a VASP OUTCAR file.
 
     Raises
@@ -232,3 +231,132 @@ def _read_lattice_from_outcar(outcar_file: TextIO) -> NDArray[np.float64]:
             raise InvalidFileException(f"lattice could not be parsed: {line}") from exc
 
     return np.array(lattice)
+
+
+def read_phonons(filepath: str | Path) -> Phonons:
+    """Extract phonons from a VASP OUTCAR file.
+
+    Parameters
+    ----------
+    filepath
+
+    Returns
+    -------
+    :
+
+    Raises
+    ------
+    FileNotFoundError
+        File not found.
+    InvalidFileException
+        File has an unexpected format.
+    """
+    wavenumbers = []
+    eigenvectors = []
+
+    filepath = pathify(filepath)
+    with open(filepath, "r", encoding="utf-8") as outcar_file:
+
+        # get atom information
+        atomic_symbols = _read_atomic_symbols(outcar_file)
+        atomic_weights = np.array([ATOMIC_WEIGHTS[symbol] for symbol in atomic_symbols])
+        num_atoms = len(atomic_symbols)
+        num_degrees_of_freedom = num_atoms * 3
+
+        # read in eigenvectors/eigenvalues
+        try:
+            _ = _skip_file_until_line_contains(
+                outcar_file, "Eigenvectors and eigenvalues of the dynamical matrix"
+            )
+        except NoMatchingLineFoundException as exc:
+            raise InvalidFileException(
+                "eigenvector/eigenvalues block not found"
+            ) from exc
+        for _ in range(num_degrees_of_freedom):
+            try:
+                line = _skip_file_until_line_contains(outcar_file, "cm-1")
+                if "f/i" in line:  # if complex
+                    wavenumbers.append(
+                        -float(line.split()[6])  # set negative wavenumber
+                    )
+                else:
+                    wavenumbers.append(float(line.split()[7]))
+            except (NoMatchingLineFoundException, TypeError, IndexError) as exc:
+                raise InvalidFileException("eigenvalue could not be parsed") from exc
+            eigenvectors.append(_read_eigenvector(outcar_file, num_atoms))
+
+        # Divide eigenvectors by sqrt(mass) to get cartesian displacements
+        wavenumbers = np.array(wavenumbers)
+        eigenvectors = np.array(eigenvectors)
+        cartesian_displacements = eigenvectors / np.sqrt(atomic_weights)[:, np.newaxis]
+
+        return Phonons(wavenumbers, cartesian_displacements)
+
+
+def read_positions_and_polarizability(
+    filepath: str | Path,
+) -> tuple[NDArray[np.float64], NDArray[np.float64]]:
+    """Extract fractional positions and polarizability from a VASP OUTCAR file.
+
+    The polarizability returned by VASP is, in fact, a dielectric tensor. However,
+    this is inconsequential to the calculation of Raman spectra.
+
+    Parameters
+    ----------
+    filepath
+
+    Returns
+    -------
+    :
+        2-tuple, whose first element is the fractional positions, a 2D array with shape
+        (N,3). The second element is the polarizability, a 2D array with shape (3,3).
+
+    Raises
+    ------
+    FileNotFoundError
+    InvalidFileException
+        If the OUTCAR has an unexpected format.
+    """
+    filepath = pathify(filepath)
+    with open(filepath, "r", encoding="utf-8") as outcar_file:
+        num_atoms = len(_read_atomic_symbols(outcar_file))
+        positions = _read_fractional_positions(outcar_file, num_atoms)
+        polarizability = _read_polarizability(outcar_file)
+        return positions, polarizability
+
+
+def read_ref_structure(
+    filepath: str | Path,
+) -> ReferenceStructure:
+    """Extract reference structure from a VASP OUTCAR file.
+
+    Parameters
+    ----------
+    filepath
+
+    Returns
+    -------
+    StructuralSymmetry
+
+    Raises
+    ------
+    InvalidFileException
+        If the OUTCAR has an unexpected format.
+    SymmetryException
+        If OUTCAR was read but the symmetry search failed
+    FileNotFoundError
+    """
+    lattice = np.array([])
+    fractional_positions = np.array([])
+    atomic_numbers = np.array([], dtype=np.int32)
+
+    filepath = pathify(filepath)
+    with open(filepath, "r", encoding="utf-8") as outcar_file:
+        atomic_symbols = _read_atomic_symbols(outcar_file)
+        atomic_numbers = np.array([ATOMIC_NUMBERS[symbol] for symbol in atomic_symbols])
+        lattice = _read_lattice(outcar_file)
+        fractional_positions = _read_fractional_positions(
+            outcar_file, len(atomic_symbols)
+        )
+
+    return ReferenceStructure(atomic_numbers, lattice, fractional_positions)
